Disabled log & hide UI option
